# openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 1.0.0
  description: OpenAPI спецификация для основного сервиса
security:
  - bearerAuth: [ ]
paths:
...
  /api/v1/main_service/projects/{projectId}/comments:
    post:
      tags:
        - ProjectComment
      summary: Создание комментария к проекту
      description: Добавляет комментарий текущего пользователя к указанному проекту.
      operationId: createProjectComment
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentFullResponse'
    get:
      tags:
        - ProjectComment
      summary: Получение комментариев проекта
      description: Возвращает список всех комментариев к проекту.
      operationId: getProjectComments
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentFullResponse'
  /api/v1/main_service/comments/{id}:
    delete:
      tags:
        - ProjectComment
      summary: Удаление комментария
      description: Удаляет комментарий с указанным идентификатором. Доступно только автору комментария.
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор комментария
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Комментарий успешно удален
          content: { }
...
  /api/v1/main_service/articles:
    post:
      tags:
        - Article
      summary: Создание статьи
      description: Создает новую статью от имени текущего пользователя
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleRequest'
      responses:
        '201':
          description: Статья была успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFullResponse'
        '400':
          description: Ошибка валидации
        '401':
          description: Неавторизованный доступ
    get:
      tags:
        - Article
      summary: Получение списка статей
      description: Возвращает список статей с фильтрацией и пагинацией
      operationId: getArticles
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Фильтр по категории
        - in: query
          name: query
          schema:
            type: string
          description: Поисковый запрос по заголовку или содержимому
        - in: query
          name: moderationPassed
          schema:
            type: boolean
          description: Фильтр по статусу модерации
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы для пагинации
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
          description: Размер страницы для пагинации
      responses:
        '200':
          description: Список статей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlePageResponse'
        '400':
          description: Ошибка параметров запроса
...
  /api/v1/main_service/articles/me:
    get:
      tags:
        - Article
      summary: Получение статей созданных текущим пользователем с пагинацией
      description: Получение статей созданных текущим пользователем с пагинацией
      operationId: getMyArticles
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Поисковый запрос по заголовку или содержимому
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы для пагинации
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
          description: Размер страницы для пагинации
      responses:
        '200':
          description: Список статей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlePageResponse'
        '400':
          description: Ошибка параметров запроса
  /api/v1/main_service/articles/{id}:
    get:
      tags:
        - Article
      summary: Получение статьи по ID
      description: Получение статьи по ID
      operationId: getArticle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Статья
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFullResponse'
        '404':
          description: Статья не найдена

    patch:
      tags:
        - Article
      summary: Обновление статьи по ID
      description: Обновление статьи по ID
      operationId: updateArticle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleRequest'
      responses:
        '200':
          description: Статья успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFullResponse'
        '404':
          description: Статья не найдена

    delete:
      tags:
        - Article
      summary: Удаление статьи по ID
      description: Удаление статьи по ID
      operationId: deleteArticle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Статья успешно удалена
        '404':
          description: Статья не найдена
  /api/v1/main_service/articles/{id}/moderation_passed:
    patch:
      tags:
        - Article
      summary: Обновление статуса модерации статьи
      description: Обновление статуса модерации статьи
      operationId: updateArticleModerationPassed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleFullResponse'
      responses:
        '200':
          description: Статус модерации успешно обновлен
        '404':
          description: Статья не найдена
...
  /api/v1/main_service/news:
    post:
      tags:
        - News
      summary: Создание новости
      description: Создание новости
      operationId: createNews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsRequest'
      responses:
        '201':
          description: Новость успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsFullResponse'
    get:
      tags:
        - News
      summary: Получение новостей с пагинацией
      description: Получение новостей с пагинацией
      operationId: getNews
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Фильтр по категории
        - name: query
          in: query
          schema:
            type: string
          description: Поисковый запрос
        - name: moderationPassed
          in: query
          schema:
            type: boolean
          description: Фильтр по статусу модерации
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Страница новостей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsPageResponse'
  /api/v1/main_service/news/me:
    get:
      tags:
        - News
      summary: Получение новостей созданных текущим пользователем с пагинацией
      description: Получение новостей созданных текущим пользователем с пагинацией
      operationId: getMyNews
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Фильтр по категории
        - name: query
          in: query
          schema:
            type: string
          description: Поисковый запрос
        - name: moderationPassed
          in: query
          schema:
            type: boolean
          description: Фильтр по статусу модерации
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Страница новостей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsPageResponse'
  /api/v1/main_service/news/{id}:
    get:
      tags:
        - News
      summary: Получение новости
      description: Получение новости
      operationId: getNewsById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Полная новость
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsFullResponse'
    patch:
      tags:
        - News
      summary: Обновление новости
      description: Обновление новости
      operationId: updateNews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsRequest'
      responses:
        '200':
          description: Новость обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsFullResponse'
    delete:
      tags:
        - News
      summary: Удаление новости
      description: Удаление новости
      operationId: deleteNews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Новость удалена
  /api/v1/main_service/news/{id}/moderation_passed:
    patch:
      tags:
        - News
      summary: Обновление статуса модерации новости
      description: Обновление статуса модерации новости
      operationId: updateNewsModerationPassed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: boolean
              example: true
      responses:
        '200':
          description: Статус модерации обновлён
components:
  schemas:
    NewsPageResponse:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/NewsSummaryResponse'
    ArticlePageResponse:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ArticleSummaryResponse'
...
    NewsRequest:
      type: object
      description: Запрос на создание или обновление новости
      properties:
        title:
          type: string
          description: Заголовок новости
        previewUrl:
          type: string
          format: uri
          description: URL превью новости
        categoryId:
          type: integer
          format: int64
          description: Категория новости
        content:
          type: string
          description: Полное содержимое новости
      required:
        - title
        - content
    NewsFullResponse:
      type: object
      description: Полный ответ по новости
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор новости
        title:
          type: string
          description: Заголовок новости
        owner:
          type: string
          description: Автор новости
        previewUrl:
          type: string
          format: uri
          description: URL превью новости
        category:
          type: string
          description: Категория новости
        content:
          type: string
          description: Полное содержимое новости
        publicationDate:
          type: string
          format: date
          description: Дата публикации новости
        moderationPassed:
          type: boolean
          description: Флаг, прошла ли новость модерацию
      required:
        - id
        - title
        - owner
        - content
        - publicationDate
        - moderationPassed
    NewsSummaryResponse:
      type: object
      description: Краткая информация о новости
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор новости
        title:
          type: string
          description: Заголовок новости
        owner:
          type: string
          description: Автор новости
        previewUrl:
          type: string
          format: uri
          description: URL превью новости
        category:
          type: string
          description: Категория новости
      required:
        - id
        - title
        - owner
    ArticleRequest:
      type: object
      description: Запрос на создание или обновление статьи
      properties:
        title:
          type: string
          description: Заголовок статьи
        previewUrl:
          type: string
          format: uri
          description: URL превью статьи
        categoryId:
          type: integer
          format: int64
          description: Категория статьи
        content:
          type: string
          description: Полное содержимое статьи
      required:
        - title
        - content
    ArticleFullResponse:
      type: object
      description: Полный ответ по статье
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор статьи
        title:
          type: string
          description: Заголовок статьи
        owner:
          type: string
          description: Владелец/автор статьи
        previewUrl:
          type: string
          format: uri
          description: URL превью статьи
        category:
          type: string
          description: Категория статьи
        content:
          type: string
          description: Полное содержимое статьи
        publicationDate:
          type: string
          format: date
          description: Дата публикации статьи
        moderationPassed:
          type: boolean
          description: Флаг, прошла ли статья модерацию
      required:
        - id
        - title
        - owner
        - previewUrl
        - category
        - content
        - publicationDate
        - moderationPassed
    ArticleSummaryResponse:
      type: object
      description: Краткая информация о статье
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор статьи
        title:
          type: string
          description: Заголовок статьи
        owner:
          type: string
          description: Владелец/автор статьи
        previewUrl:
          type: string
          format: uri
          description: URL превью статьи
        category:
          type: string
          description: Категория статьи
      required:
        - id
        - title
        - owner
        - previewUrl
        - category
...
    CommentRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Текст комментария
          minLength: 1
          example: "test"
    Page:
      type: object
      required:
        - content
        - number
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: Массив элементов текущей страницы
          items: { }
        number:
          type: integer
          example: 1
        size:
          type: integer
          example: 1
        totalElements:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1
    UserFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 42
        username:
          type: string
          description: Имя пользователя
          example: "Andrey"
...
    CommentFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор комментария
          example: 1001
        owner:
          $ref: '#/components/schemas/UserFullResponse'
        text:
          type: string
          description: Текст комментария
          example: "User12345"
        createdDate:
          type: string
          format: date-time
          description: Дата и время создания комментария (UTC)
          example: "2026-01-24T14:32:10"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT